Source files detached and fullfilled with code
