+++ include/task1_graph/drawer/graph_drawer.h
@@ -13,7 +13,7 @@ class graphDrawer
 
     sf::Font font;
 
-    graphDrawer() = default;
+    graphDrawer();
 
     ~graphDrawer() = default;
 
@@ -34,151 +34,55 @@ public:
 #pragma region window_and_rendering_utilities
 
 public:
-    void create_window()
-    {
-        if (window != nullptr)
-            window = nullptr;
+    void create_window();
 
-        window = new sf::RenderWindow({1280u, 1024u}, "Main Window");
-        window->setFramerateLimit(120u);
-        font.loadFromFile("arial.ttf");
-    }
-
-    void render()
-    {
-        if (window == nullptr) return;
-
-        listen();
-        clear();
-        display();
-    }
+    void render();
 
 private:
-    void listen()
-    {
-        if (window->isOpen())
-        {
-            for (auto event = sf::Event(); window->pollEvent(event);)
-            {
-                if (event.type == sf::Event::Closed)
-                {
-                    shutdownFlag = true;
-                    window->close();
-                    break;
-                }
-            }
-        }
-    }
+    void listen();
 
-    void display() const
-    {
-        for (const auto& [key, value] : frame.get_drawn_context())
-            window->draw(*value);
+    void display() const;
 
-        window->display();
-    }
-
-    void clear() const
-    {
-        window->clear();
-    }
+    void clear() const;
 
 #pragma endregion window_and_rendering_utilities
 
 #pragma region drawing_api
-#define N_ARROW_SHAPE_POINTS 7
-#define ARROW_HEAD_SHIFT 6.0f
-#define FONT_SIZE 14
 
 public:
-    unsigned int draw_circle(const float radius,
-                             const float x,
-                             const float y,
-                             const sf::Color& color)
-    {
-        const auto circle = new sf::CircleShape(radius);
-        circle->setPosition(x, y);
-        circle->setFillColor(color);
-        return frame.push(circle);
-    }
+    unsigned int draw_circle(float radius,
+                             float x,
+                             float y,
+                             const sf::Color& color = sf::Color::Yellow);
 
-    unsigned int draw_line(const float length,
-                           const float width,
-                           const float x,
-                           const float y,
-                           const float rotation,
-                           const sf::Color& color)
-    {
-        const auto line = new sf::RectangleShape({length, width});
-        line->setPosition({x, y});
-        line->setRotation(rotation);
-        line->setFillColor(color);
-        return frame.push(line);
-    }
+    unsigned int draw_line(float length,
+                           float width,
+                           float x,
+                           float y,
+                           float rotation,
+                           const sf::Color& color = sf::Color::Blue);
 
-    unsigned int draw_arrow(const float xStart,
-                            const float yStart,
-                            const float xEnd,
-                            const float yEnd,
-                            const float width,
-                            const sf::Color& color)
-    {
-        const auto arrow = new sf::ConvexShape();
-        arrow->setPointCount(N_ARROW_SHAPE_POINTS);
-        // TODO: Сформировать фигуру для рендера стрелки
-        arrow->setPoint(0, {xStart - width / 2, yStart - width / 2});
-        arrow->setPoint(1, {xStart + width / 2, yStart + width / 2});
-        arrow->setPoint(2, {xStart - width / 2, yStart - width / 2});
-        arrow->setPoint(3, {xEnd - width / 2, yEnd - width / 2});
-        arrow->setPoint(4, {xEnd - width / 2, yEnd - width / 2});
-        arrow->setPoint(5, {xEnd - width / 2, yEnd - width / 2});
-        arrow->setPoint(6, {xEnd - width / 2, yEnd - width / 2});
-        arrow->setFillColor(color);
-        return frame.push(arrow);
-    }
+    unsigned int draw_arrow(float xStart,
+                            float yStart,
+                            float xEnd,
+                            float yEnd,
+                            float width,
+                            const sf::Color& color = sf::Color::Blue);
 
     unsigned int draw_label(const std::string& label,
-                            const float x,
-                            const float y,
-                            const sf::Color& color)
-    {
-        const auto text = new sf::Text(label, font, FONT_SIZE);
-        text->setPosition(x, y);
-        text->setFillColor(color);
-        return frame.push(text);
-    }
+                            float x,
+                            float y,
+                            const sf::Color& color = sf::Color::Magenta);
 
-    void erase_drawing(const unsigned int id)
-    {
-        const auto drawing = frame.get_drawable(id);
-        frame.erase(id);
-        delete drawing;
-    }
+    void erase_drawing(unsigned int id);
 
-    void recolor_label(const unsigned int id, const sf::Color& color)
-    {
-        frame.get_text(id)->setFillColor(color);
-    }
+    void recolor_label(unsigned int id, const sf::Color& color);
 
-    void recolor_shape(const unsigned int id, const sf::Color& color)
-    {
-        frame.get_shape(id)->setFillColor(color);
-    }
+    void recolor_shape(unsigned int id, const sf::Color& color);
 
-    void rename_label(const unsigned int id, const std::string& name)
-    {
-        const auto text = frame.get_text(id);
-        const auto previous_name = text->getString();
-        delete previous_name.getData();
-        // TODO: Check if this memory management is correct for SFML sf::String
-        text->setString(name);
-    }
+    void rename_label(unsigned int id, const std::string& name);
 
-    void replace_shape(const unsigned int id,
-                       const float x,
-                       const float y)
-    {
-        frame.get_shape(id)->setPosition(x, y);
-    }
+    void replace_shape(unsigned int id, float x, float y);
+
 #pragma endregion drawing_api
 };
