--- src/drawer_command.cpp
@@ -1,3 +1,0 @@
-//
-// Created by AleksandrIvanitskii on 11/14/2024.
-//
